Rename utils and split into describe & dimensional
