+++ python/dlisio/plumbing/message.py
@@ -1,6 +1,6 @@
 from .basicobject import BasicObject
 from .valuetypes import scalar, vector
-from .utils import describe_array, describe_attributes
+from .describe import describe_array, describe_attributes
 
 from collections import OrderedDict
